Improvments regarding reusability of the common role
Here the common role is plan to be used for all applications that
require only one package and where all the configuration is in
$XDG_CONFIG_HOME.

The loop on all items of the applications dictionnary is done in the
playbooks by including the role common to a task where the loop is
done.

I have still to ensure that uncommon applications are simply skipped
in the playbook and not in the role.

A few modifications have been done in config-xdg and in the
configuration template to better handle user environement creation.

Still a lot of work to do.